// front/src/helpers/types.tsx
export interface SouFormData {
  id?: string;
  userId?: string;
  nome: string;
  idade: string;
  dataNascimento: string;
  curso: string;
  fase: string;
  turno: string;
  dificuldade: string;
  disciplina: boolean;
  explicacao: boolean;
  sala: boolean;
  colegas: boolean;
  preferencia: string;
  horario: string;
  data: string;
  observacoes: string;
  profissional: string;
}